WIP BASE CODE

+++ src/Pin/Pin.ino
@@ -1,9 +1,10 @@
+#include <Arduino.h>
+
 #include "Pin.h"
 
-Pin::Pin(String name, uint8_t number, PinType type) 
+Pin::Pin(String name, int number, PinType type)
 {
   this->name = name;
   this->number = number;
   this->type = type;
 }
-
